requirements.txt
Django==2.2.4
google-images-download==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
line-bot-sdk==1.13.0
Markdown==3.1.1
numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
